Edit Users and roles and admin control

+++ resources/views/admin/users/index.blade.php
@@ -28,13 +28,24 @@
                             <td>{{ $user -> email }}</td>
                             <td>{{ DB::table('roles')->where('id', $user->role_id )->value('name') }}</td>
                             <td>
-                            {{--<a href="#"><button type="button" class="btn btn-success"><i class="fas fa-edit"></i></button></a>--}}
-                            {!! Form::open(array('route'=>['users.edit',$user->id],'method'=>'EDIT')) !!}
-                            {!! Form::button('<i class="fas fa-edit"></i>',['class'=>'btn btn-success','type'=>'submit']) !!}
-                            {!! Form::close() !!}
-                            {!! Form::open(array('route'=>['users.destroy',$user->id],'method'=>'DELETE')) !!}
-                            {!! Form::button('<i class="fas fa-trash-alt"></i>',['class'=>'btn btn-danger','type'=>'submit']) !!}
-                            {!! Form::close() !!}
+                                @csrf
+                                @if(Auth::user()->role_id == 1 || ($user->role_id ==Auth::user()->role_id && $user->id==Auth::user()->id))
+                                    {{--<a href="#"><button type="button" class="btn btn-success"><i class="fas fa-edit"></i></button></a>--}}
+                                    <div style="float:left;">
+                                        {!! Form::open(array('route'=>['users.edit',$user->id], 'method'=>'GET')) !!}
+                                        {!! Form::button('<i class="fas fa-edit"></i>',['class'=>'btn btn-success','type'=>'submit', 'style'=>'margin-right:-75px;']) !!}
+                                        {!! Form::close() !!}
+                                    </div>
+                                    {{--</td>--}}
+                                    {{--<td>--}}
+                                    <div style="float:left; margin-left: 45px;">
+                                        {!! Form::open(array('route'=>['users.destroy',$user->id],'method'=>'DELETE')) !!}
+                                        {!! Form::button('<i class="fas fa-trash-alt"></i>',['class'=>'btn btn-danger','type'=>'submit']) !!}
+                                        {!! Form::close() !!}
+                                    </div>
+                                @endif
+                            </td>
+
                         </tr>
                     @endforeach
                     </tbody>
